add text input and make it connected to display box using onchange
